# Swagger/booking-openapi.yaml
openapi: "3.1.0"
info:
  title: "booking API"
  description: "booking API"
  version: "1.0.0"
servers:
  - url: "https://booking"
paths:
  /api/manager/hotels:
    get:
      summary: "GET api/manager/hotels"
      operationId: "listHotels"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Hotel"
  /api/manager/hotels/{id}:
    get:
      summary: "GET api/manager/hotels/{id}"
      operationId: "getHotelById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Hotel"
    delete:
      summary: "DELETE api/manager/hotels/{id}"
      operationId: "deleteHotel"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
    put:
      summary: "PUT api/manager/hotels/{id}"
      operationId: "editHotel"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Hotel"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Hotel"
  /api/manager/hotels/add:
    post:
      summary: "POST api/manager/hotels/add"
      operationId: "addHotel"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Hotel"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Hotel"
  /api/manager/hotels/{hotelId}/rooms:
    get:
      summary: "GET api/manager/hotels/{hotelId}/rooms"
      operationId: "listRooms"
      parameters:
        - name: "hotelId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Room"
  /api/manager/hotels/{hotelId}/rooms/{roomId}:
    get:
      summary: "GET api/manager/hotels/{hotelId}/rooms/{roomId}"
      operationId: "getRoomById"
      parameters:
        - name: "hotelId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
        - name: "roomId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      summary: "DELETE api/manager/hotels/{hotelId}/rooms/{roomId}"
      operationId: "deleteRoom"
      parameters:
        - name: "hotelId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
        - name: "roomId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
  /api/manager/hotels/{hotelId}/rooms/add:
    post:
      summary: "POST api/manager/hotels/{hotelId}/rooms/add"
      operationId: "addRoom"
      parameters:
        - name: "hotelId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Room"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Room"
  /api/manager/hotels/{hotelId}/rooms/{id}:
    put:
      summary: "PUT api/manager/hotels/{hotelId}/rooms/{id}"
      operationId: "editRoom"
      parameters:
        - name: "hotelId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Room"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Room"
components:
  schemas:
    Role:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        roleType:
          type: "string"
          enum:
            - "CLIENT"
            - "HOTEL_MANAGER"
    BigDecimal:
      type: "object"
      properties: { }
    Unavailability:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        startDate:
          type: "string"
          format: "date-time"
        endDate:
          type: "string"
          format: "date-time"
        room:
          $ref: "#/components/schemas/Room"
    Room:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        description:
          type: "string"
        currentPrice:
          $ref: "#/components/schemas/BigDecimal"
        roomType:
          type: "string"
          enum:
            - "SINGLE"
            - "DOUBLE"
        hotel:
          $ref: "#/components/schemas/Hotel"
        bookingList:
          type: "array"
          items:
            $ref: "#/components/schemas/Booking"
        unavailabilityList:
          type: "array"
          items:
            $ref: "#/components/schemas/Unavailability"
    Booking:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        startDate:
          type: "string"
          format: "date-time"
        endDate:
          type: "string"
          format: "date-time"
        created:
          type: "string"
          format: "date-time"
        updated:
          type: "string"
          format: "date-time"
        discountPercent:
          $ref: "#/components/schemas/BigDecimal"
        bookingStatus:
          type: "string"
          enum:
            - "CONFIRMED"
            - "CANCELLED"
        totalPrice:
          $ref: "#/components/schemas/BigDecimal"
        client:
          $ref: "#/components/schemas/Client"
        room:
          $ref: "#/components/schemas/Room"
    Client:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        user:
          $ref: "#/components/schemas/User"
        bookingList:
          type: "array"
          items:
            $ref: "#/components/schemas/Booking"
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        username:
          type: "string"
        password:
          type: "string"
        createdDate:
          type: "string"
          format: "date-time"
        role:
          $ref: "#/components/schemas/Role"
        name:
          type: "string"
        lastName:
          type: "string"
        hotelManager:
          $ref: "#/components/schemas/HotelManager"
        client:
          $ref: "#/components/schemas/Client"
    HotelManager:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        user:
          $ref: "#/components/schemas/User"
        hotelList:
          type: "array"
          items:
            $ref: "#/components/schemas/Hotel"
    Address:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        line:
          type: "string"
        city:
          type: "string"
        country:
          type: "string"
        hotel:
          $ref: "#/components/schemas/Hotel"
    Hotel:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        description:
          type: "string"
        categoryType:
          type: "string"
          enum:
            - "HOTEL"
        hotelManager:
          $ref: "#/components/schemas/HotelManager"
        roomList:
          type: "array"
          items:
            $ref: "#/components/schemas/Room"
        address:
          $ref: "#/components/schemas/Address"